+++ docs/changes.rst
@@ -18,6 +18,7 @@ v2.3.1 (UNRELEASED)
 - Update: ``poledancingadventures`` after site change.
 - Update: ``pcweenies`` after feed change.
 - Update: ``yamac`` after site change.
+- Inactive: ``antics``
 - Inactive: ``boxerhockey``
 - Inactive: ``eatthattoast``
 - Inactive: ``gregcomic``